+++ src/camera.h
@@ -11,30 +11,36 @@ class Camera
 {    
 public:
 
-	Camera(const Vector3d &pos, const Vector3d &dir, const Vector3d &up, const double fov, const double aspect_ratio,
+	Camera(const Vector3f &pos, const Vector3f &dir, const Vector3f &up, const double fov, const double aspect_ratio,
 		const Vector2i &resolution, double focal_length, double blur_radius);
 
-	void set_position(const Vector3d &pos) {
+	void set_position(const Vector3f &pos) 
+	{
 		m_position = pos;
 	}
 
-	const Vector3d position() const {
+	const Vector3f position() const
+	{
 		return m_position;
 	}
 
-	void set_direction(const Vector3d &dir) {
+	void set_direction(const Vector3f &dir) 
+	{
 		m_direction = dir;
 	}
 
-	const Vector3d direction() const {
+	const Vector3f direction() const 
+	{
 		return m_direction;
 	}
 
-	double blur_radius() const {
+	double blur_radius() const 
+	{
 		return m_blur_radius;
 	}
 
-	void set_blur_radius(const double value) {
+	void set_blur_radius(const double value) 
+	{
 		m_blur_radius = value;
 	}
 
@@ -42,11 +48,13 @@ public:
 		return m_focal_length;
 	}
 
-	void set_focal_length(double value) {
+	void set_focal_length(double value) 
+	{
 		m_focal_length = value;
 	}
 
-	bool updated_this_frame() const {
+	bool updated_this_frame() const 
+	{
 		return m_updated_this_frame;
 	}
 
@@ -58,12 +66,12 @@ public:
 
 	void reset_update_flag();
 private:
-    Vector3d m_position;
-	Vector3d m_direction;
-	Vector3d m_up;
-	Vector3d m_u;
-	Vector3d m_v;
-    Vector3d m_image_plane_origin;
+    Vector3f m_position;
+	Vector3f m_direction;
+	Vector3f m_up;
+	Vector3f m_u;
+	Vector3f m_v;
+    Vector3f m_image_plane_origin;
 	double m_blur_radius;
 	double m_fov;
 	double m_aspect_ratio;
