+++ src/mesh.h
@@ -15,14 +15,14 @@ class Mesh
 {
 public:
 	
-	std::vector<Vector3d> m_vertices;
+	std::vector<Vector3f> m_vertices;
 	std::vector<uint16_t> m_indices;
-	std::vector<Vector3d> m_normals;
-	std::vector<Vector2d> m_texture_coords;
+	std::vector<Vector3f> m_normals;
+	std::vector<Vector2f> m_texture_coords;
 	std::shared_ptr<Material> m_material;
 	AABB m_bounds;
 
-	Mesh(const std::string &name, const std::vector<Vector3d> &vertices, const std::vector<uint16_t> &indices, const std::vector<Vector3d> &normals, const std::vector<Vector2d> &texture_coords, const std::shared_ptr<Material> &material);
+	Mesh(const std::string &name, const std::vector<Vector3f> &vertices, const std::vector<uint16_t> &indices, const std::vector<Vector3f> &normals, const std::vector<Vector2f> &texture_coords, const std::shared_ptr<Material> &material);
 
 	void transform_vertices(const Matrix &transform, const Matrix &normal_transform);
 
@@ -31,8 +31,8 @@ public:
 
 	const AABB &bounds() const { return m_bounds; }
 private:
-	std::vector<Vector3d> m_original_vertices;
-	std::vector<Vector3d> m_original_normals;
+	std::vector<Vector3f> m_original_vertices;
+	std::vector<Vector3f> m_original_normals;
     std::string m_name;
 	void update_bounds();
 };
